未ログイン時や本人以外のアクセスの制限設定

+++ app/controllers/studies_controller.rb
@@ -1,4 +1,5 @@
 class StudiesController < ApplicationController
+  before_action :authenticate_user!, only: [:new, :edit, :show]
   before_action :set_room
   before_action :set_study, only: [:edit, :update, :show]
 
@@ -19,20 +20,31 @@ class StudiesController < ApplicationController
   end
 
   def edit
+    unless current_user == @study.user
+      redirect_to study_path(@study)
+    end
   end
 
   def update
-    if @study.update(study_params)
-      redirect_to room_room_users_path(@room)
+    if current_user == @study.user
+      if @study.update(study_params)
+        redirect_to room_room_users_path(@room)
+      else
+        render :edit, status: :unprocessable_entity
+      end
     else
-      render :edit, status: :unprocessable_entity
+      redirect_to study_path(@study)
     end
   end
 
   def destroy
-    study = Study.find(params[:id])
-    study.destroy
-    redirect_to room_room_users_path(@room)
+    if current_user == @study.user
+      study = Study.find(params[:id])
+      study.destroy
+      redirect_to user_path(@study.user)
+    else
+      redirect_to study_path(@study)
+    end
   end
 
   private
